Rename IReferenceUnboxingProjection to IReferenceToOptionalProjection and make it directly generic.
